chore: move `PositiveNonZeroInt` to types
