+++ crates/css_ast/src/types/mod.rs
@@ -46,6 +46,7 @@ mod paint_box;
 mod palette_identifier;
 mod position;
 mod position_area;
+mod positive_non_zero_int;
 mod quote;
 mod ratio;
 mod relative_size;
@@ -126,6 +127,7 @@ pub use paint_box::*;
 pub use palette_identifier::*;
 pub use position::*;
 pub use position_area::*;
+pub use positive_non_zero_int::*;
 pub use quote::*;
 pub use ratio::*;
 pub use relative_size::*;
